started separating endurance, station keeping, fleet race into sep files
